[♻️ refactor/#101]: SignUpService 클래스 제거 및 AuthService에 회원가입 로직 통합
- `SignUpService` 클래스를 제거하고, 해당 로직을 `AuthService`로 통합하였습니다.
- `AuthController`에서 `SignUpService`를 호출하던 부분을 `AuthService` 호출로 수정하였습니다.
- 코드 중복을 줄이고, 회원가입 로직을 `AuthService`로 일원화하여 서비스 구조를 단순화하였습니다.
